requirements.txt
streamlit==1.31.0
pandas==2.1.0
numpy==1.26.0
Pillow==10.0.0
face-recognition==1.3.0
opencv-python==4.8.0.76 